recon_rec (recon v2.5.3)

This module handles formatting records for known record types. Record definitions are imported from modules by user. Definitions are distinguished by record name and its arity, if you have multiple records of the same name and size, you have to choose one of them and some of your records may be wrongly labelled. You can manipulate your definition list by using import/1 and clear/1, and check which definitions are in use by executing list/0.

Types

field/0

-type field() :: atom().

import_result/0

-type import_result() ::
    {imported, module(), record_name(), arity()} |
    {overwritten, module(), record_name(), arity()} |
    {ignored, module(), record_name(), arity(), module()}.

limit/0

-type limit() :: all | none | field() | [field()].

listentry/0

-type listentry() :: {module(), record_name(), [field()], limit()}.

record_name/0

-type record_name() :: atom().

compound

Functions

clear()

remove all imported definitions, destroy the table, clean up

clear(Module)

remove definitions imported from a module.

get_list()

-spec get_list() -> [listentry()].

returns a list of active record definitions

import(Modules)

-spec import(module() | [module()]) -> import_result() | [import_result()].

import record definitions from a module. If a record definition of the same name and arity has already been imported from another module then the new definition is ignored (returned info tells you from which module the existing definition was imported). You have to choose one and possibly remove the old one using clear/1. Supports importing multiple modules at once (by giving a list of atoms as an argument).

is_active()

-spec is_active() -> boolean().

quickly check if we want to do any record formatting

limit(Name, Arity, Limit)

-spec limit(record_name(), arity(), limit()) -> ok | {error, any()}.

Limit output to selected fields of a record (can be 'none', 'all', a field or a list of fields). Limit set to 'none' means there is no limit, and all fields are displayed; limit 'all' means that all fields are squashed and only record name will be shown.

list()

prints out all "known" (imported) record definitions and their limit settings. Printout tells module a record originates from, its name and a list of field names, plus the record's arity (may be handy if handling big records) and a list of field it limits its output to, if set.