recon_lib (Recon v2.5.1)

Regroups useful functionality used by recon when dealing with data from the node. The functions in this module allow quick runtime access to fancier behaviour than what would be done using recon module itself.

Types

diff()

Specs

t:recon_lib.diff/0

Functions

count(Terms)

Specs

recon_lib.count/1

Takes a list of terms, and counts how often each of them appears in the list. The list returned is in no particular order.

inet_attrs(AttrName)

Specs

recon_lib.inet_attrs/1

Returns the attributes (recon:inet_attrs()) of all inet ports (UDP, SCTP, TCP) of the node.

inet_attrs(AttributeName, Port)

Specs

recon_lib.inet_attrs/2

Returns the attributes required for a given inet port (UDP, SCTP, TCP). This form of attributes is standard for most comparison functions for processes in recon.

insert(E, H)

merge(H1, H2)

merge_pairs(T)

pheap_fill(List, N, Heap)

pheap_full(T, Heap)

pheap_to_list(T, Acc)

port_list(Attr)

Specs

recon_lib.port_list/1

Returns a list of all the open ports in the VM, coupled with one of the properties desired from erlang:port_info/1-2.

port_list(Attr, Val)

Specs

recon_lib.port_list/2

Returns a list of all the open ports in the VM, but only if the Attr's resulting value matches Val. Attr must be a property accepted by erlang:port_info/2.

proc_attrs(AttrName)

Specs

recon_lib.proc_attrs/1

Returns the attributes (recon:proc_attrs()) of all processes of the node, except the caller.

proc_attrs(AttrName, Pid)

Specs

recon_lib.proc_attrs/2

Returns the attributes of a given process. This form of attributes is standard for most comparison functions for processes in recon.
A special attribute is binary_memory, which will reduce the memory used by the process for binary data on the global heap.

sample(Ms, Fun)

Specs

recon_lib.sample/2

Runs a fun once, waits Ms, runs the fun again, and returns both results.

scheduler_usage_diff(SchedTime, SchedTime)

Specs

recon_lib.scheduler_usage_diff/2

Diffs two runs of erlang:statistics(scheduler_wall_time) and returns usage metrics in terms of cores and 0..1 percentages.

sliding_window(First, Last)

Specs

recon_lib.sliding_window/2

Compare two samples and return a list based on some key. The type mentioned for the structure is diff() ({Key,Val,Other}), which is compatible with the recon:proc_attrs() type.

sublist_top_n_attrs(List, Len)

Specs

recon_lib.sublist_top_n_attrs/2

Returns the top n element of a list of process or inet attributes

term_to_pid(Pid)

Specs

recon_lib.term_to_pid/1

Transforms a given term to a pid.

term_to_port(Port)

Specs

recon_lib.term_to_port/1

Transforms a given term to a port

time_fold(N, Interval, Fun, State, FoldFun, Init)

Specs

recon_lib.time_fold/6

Calls a given function every Interval milliseconds and supports a fold-like interface (each result is modified and accumulated)

time_map(N, Interval, Fun, State, MapFun)

Specs

recon_lib.time_map/5

Calls a given function every Interval milliseconds and supports a map-like interface (each result is modified and returned)

triple_to_pid(N, N, N)

Specs

recon_lib.triple_to_pid/3

Equivalent of pid(X,Y,Z) in the Erlang shell.