recon_map (Recon v2.5.1)

This module handles formatting maps. It allows for trimming output to selected fields, or to nothing at all. It also adds a label to a printout. To set up a limit for a map, you need to give recon a way to tell the map you want to trim from all the other maps, so you have to provide something like a 'type definition'. It can be either another map which is compared to the arg, or a fun.

Types

limit()

Specs

t:recon_map.limit/0

map_label()

Specs

t:recon_map.map_label/0

pattern()

Specs

t:recon_map.pattern/0

Functions

apply_map_limits(Fields, M)

clear()

remove all imported definitions, destroy the table, clean up

ensure_table_exists()

ets_keeper()

is_active()

Specs

recon_map.is_active/0

quickly check if we want to do any record formatting

limit(Label, Pattern, Limit)

Specs

recon_map.limit/3

Limit output to selected keys of a map (can be 'none', 'all', a key or a list of keys). Pattern selects maps to process: a "pattern" is just a map, and if all key/value pairs of a pattern are present in a map (in other words, the pattern is a subset), then we say the map matches and we process it accordingly (apply the limit).

Patterns are applied in alphabetical order, until a match is found.
Instead of a pattern you can also provide a function which will take a map and return a boolean.

list()

prints out all "known" map definitions and their limit settings. Printout tells a map's name, the matching fields required, and the limit options.

list(Rest)

prints out all "known" map filter definitions and their settings. Printout tells the map's label, the matching patterns, and the limit options

map_matches(M, Pattern)

maybe_kill(Name)

patterns_table_name()

prepare_limit(Limit)

prepare_pattern(Pattern)

process_map(M, Rest)

remove(Label)

Specs

recon_map.remove/1

remove a given map entry

rename(Name, NewName)

Specs

recon_map.rename/2

rename a given map entry, which allows to to change priorities for matching. The first argument is the current name, and the second argument is the new name.

store_pattern(Label, Pattern, Limit)

wait_for_death(Pid, Name)