recon_rec (Recon v2.5.1)

This module handles formatting records for known record types. Record definitions are imported from modules by user. Definitions are distinguished by record name and its arity, if you have multiple records of the same name and size, you have to choose one of them and some of your records may be wrongly labelled. You can manipulate your definition list by using import/1 and clear/1, and check which definitions are in use by executing list/0.

Types

field()

Specs

t:recon_rec.field/0

import_result()

Specs

t:recon_rec.import_result/0

limit()

Specs

t:recon_rec.limit/0

listentry()

Specs

t:recon_rec.listentry/0

record_name()

Specs

t:recon_rec.record_name/0

compound

Functions

apply_limits(List, Field)

clear()

remove all imported definitions, destroy the table, clean up

clear(Module)

remove definitions imported from a module.

ets_keeper()

field_name(_)

field_names(Fields)

format_default(Val)

format_kv(Key, Val)

format_record(Rec, _)

format_tuple(Name, Rec)

get_list()

Specs

recon_rec.get_list/0

returns a list of active record definitions

get_record(_, Acc)

get_record_defs(Module)

import(Modules)

Specs

recon_rec.import/1

import record definitions from a module. If a record definition of the same name and arity has already been imported from another module then the new definition is ignored (returned info tells you from which module the existing definition was imported). You have to choose one and possibly remove the old one using clear/1. Supports importing multiple modules at once (by giving a list of atoms as an argument).

import(Module, ResultList)

is_active()

Specs

recon_rec.is_active/0

quickly check if we want to do any record formatting

join(Sep, List)

Specs

recon_rec.join/2

limit(Name, Arity, Limit)

Specs

recon_rec.limit/3

Limit output to selected fields of a record (can be 'none', 'all', a field or a list of fields). Limit set to 'none' means there is no limit, and all fields are displayed; limit 'all' means that all fields are squashed and only record name will be shown.

list()

prints out all "known" (imported) record definitions and their limit settings. Printout tells module a record originates from, its name and a list of field names, plus the record's arity (may be handy if handling big records) and a list of field it limits its output to, if set.

make_list_entry(_)

maybe_kill(Name)

rec_info(_, Module)

records_table_name()

rem_for_module(Rec, Module)

store_record(Rec, Module, ResultList)

wait_for_death(Pid, Name)